:art: Organise code.
