+++ src/snake_ai_pytorch/controllers/play.py
@@ -1,0 +1,5 @@
+from snake_ai_pytorch.models import SnakeGameHuman
+
+if __name__ == "__main__":
+    env = SnakeGameHuman()
+    env.run()
