+++ src/snake_ai_pytorch/controllers/train.py
@@ -1,0 +1,7 @@
+from snake_ai_pytorch.models import Agent
+
+if __name__ == "__main__":
+    # agent = Agent()
+    agent = Agent(render_mode="human")
+    agent = Agent(render_mode="fast_training")
+    agent.train()
