+++ src/snake_ai_pytorch/views/__init__.py
@@ -1,1 +1,2 @@
+from snake_ai_pytorch.views.progress_graph import plot  # noqa: F401
 from snake_ai_pytorch.views.visual_configuration import BLOCK_SIZE, SPEED, FontConfig, GameColors  # noqa: F401
